Changed method for asserting the result
